🔎 Fix lookup of embedded resources
Now that the root namespace for all assemblies is the same,
the base name for the embedded resource no longer matches
the assembly name. So in order to avoid this and future issues
with other generators, just make the search more flexible
by finding the resource name by relative path name instead.