🔗 Support linked files in FileConstant
It's not uncommon to include files from outside the current project cone by linking them so they get properly copied to a desired output location (or shown in the project structure properly). If a `%(Link)` is present, use that instead of `RelativeDir` to determine the constant area and value to generate.

Fixes #38