Make the C# language check a warning instead of an error
Instead of reporting an error (both on BeforeCompile and CoreCompile) when we detect the project language isn't C# 9.0+, which breaks design-time builds (in addition to full builds) and causes bad IDE experience right-after install, make it a warning instead.

This gives the user the chance to build, perhaps attempt to type ThisAssembly and see that nothing happens, note the warning, and then cleanly uninstall the package.

Fixes #51