Add ThisAssembly.Git, leveraging Microsoft.SourceLink and ThisAssembly.Constants
This new package is just targets that leverage git-based SourceLink
packages (which must be installed separately since there is a variety)
to populate source control information, turn it into constants that then
the ThisAssembly.Constants package surfaces in a standard manner.

We do not force run of source control queries in design-time builds,
unless the user forces it. This keeps the performance characteristics
of SourceLink, at the expense of a slightly confusing user experience
where intellisense might show the properties as empty all the time
while in fact it's just during design-time.

Closes #69