Fix build issue when string resource has newlines
The value when no Comment is provided in teh resource file is used to make up the code comment for the generated property. This breaks if the value has newlines in it.

A workaround is for users to provide a shorter description in the resx, but we now also sanitize the value by replacing newlines with a whitespace.