Bump to latest SponsorLink and ensure private assets always
Since analyzers are transitive by default (unless PrivateAssets=all in the PackageReference), we could have been inadvertently generating ThisAssembly classes via project-to-project (P2P) references and causing potential type name conflicts (if InternalsVisibleTo is also in use).

This can happen if you add a package reference manually and forget to set PrivateAssets=all. But ThisAssembly is *always* intended to be private assets.

After some testing, it turns out that even if we set the PrivateAssets=all via targets provided by the package itself, we can prevent this abnormal (but perhaps easy to encounter?) situation from happening at all.

This is particularly necessary with the introduction of the SponsorLink checks, since those are analyzers too, and require compiler-visible properties to be surfaced.

By forcing PrivateAssets, we make sure SponsorLink analyzer never runs on P2P projects and only on the directly referencing one.