Only run targets when necessary and no earlier
When we add items/propeties that need to be visible to the generators, we only really need to run before the editorconfig is emitted, and no earlier. This happens right before compilation, so it's a safe late-stage target to depend on.

We cannot run later since that would mean outdated info to the generator, but this should avoid early-run issues such as https://github.com/devlooped/GitInfo/issues/260.