Allow using ThisAssembly.Resources and ThisAssembly.Strings SxS
This fixes an issue where the additional file metadata (as persisted by the generated editorconfig for that) is duplicated and overriden between these two packages since both use EmbeddedResource item to drive the source generator.

Switch to per-generator item metadata and avoid using AdditionalFiles item metadata, which has this problem.