Unify and future-proof Assembly*Attributes with Constants
One more unification and this is likely the last one.
Rather than a custom generator for assembly info attributes, we leverage MSBuild (similar to how the Metadata package does it) to transform the same set of attributes and turn them into @(Constant) items.

Now that custom and dynamic roots are possible, this is really straightforward and should be 100% backwards compatible.

In the future, if additional aseembly-level attributes matching the `Assembly[Name]Attribute` convention are added to the SDK, they will be surfaced automatically as well.