Allow setting visibility and switching to static readonly props
Setting the `ThisAssembly` class to public should be somewhat rare, but might be desirable in some cases. When the class is made public, there's a surprising "bug" in that updating the assembly providing the `ThisAssembly` class will *not* cause another assembly consuming its values to reflect the changes. This is because the compiler will efectively inline the constants, so effectively every call site has a copy of the value.

This is somewhat unintuitive, but it's perfectly sensible when the use is for internal purposes.

The new ThisAssemblyVisibility=public property allows both setting the generated class visibility to public and switching (for constants) to static properties.

For ThisAssembly.Strings and ThisAssembly.Resources, only the class visibility is changed.

Fixes #64