Avoid non-Git constants conflict, update root comment
As presently defined, if a custom constant in user's project had a naming conflict (i.e. `Root` or `Url`), the item metadata in the editorconfig/options would be mixed up, since it uses the item spec for that.

By using the full area name, we avoid that, and switch instead to the custom root feature of using `Root=.` instead.

We also take advantage of the new custom root summary comment to better reflect the nature of the provided constants.